:recycle: Rename existing transaction pool to be deleted later
