+++ elements/lisk-transaction-pool/test/integration/transaction_pool_events.spec.ts
@@ -17,7 +17,7 @@ import {
 	TransactionPool,
 	TransactionPoolConfiguration,
 	EVENT_VERIFIED_TRANSACTION_ONCE,
-} from '../../src/transaction_pool';
+} from '../../src/legacy_transaction_pool';
 import { Transaction } from '../../src/types';
 import {
 	fakeCheckFunctionGenerator,
